UBUNTU: try to automatically get version to give for dpkg-genchanges
dpkg-genchanges includes the changelog only for the last version if no
version is specified to dpkg-buildpackage. The problem is that sometimes
we want to include the changes up to more previous versions, like when
packages have some version in -proposed and another version bump happens
before it goes to -updates, and we need the changelog to contain more
entries up to the previous version in -updates.

To solve this, we try now automatically to get the previous version that
was previously released either to updates or release pockets, using
rmadison. And if still someone wants to override that, LAST_VERSION
variable can be specified on the command line that it will be used with
dpkg-buildpackage instead of using rmadison.

While at it, also get the package name from the changelog and use it
throughout the Makefile instead of hard coding it.

Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>