gator-daemon: Fix sequence-point compile error
In mxml/mxml-file.c in the expression:
  (getc(fp) << 8) | getc(fp)
the C standard allows right-hand getc() to be called before left hand
one, which would cause a bug.

The compiler gives an warning about this and we are treating warnings as
errors this breaks the build with:
  error: operation on 'fp->_r' may be undefined [-Werror=sequence-point]

We can fix this by rewriting the expressions in two parts, however the
code in question has licensing clauses for modified versions which have
not yet been resolved.

Therefore, for now, we will disable the problematic error in the
makefiles.

Signed-off-by: Jon Medhurst <tixy@linaro.org>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>