UBUNTU: Move dependency on crda to extra package
With the split between linux-image and linux-image-extra, all
wireless modules are in the extra package. Only some base
libraries were still in the main package. Move those into
extra and also move the dependency on crda/wireless-tools
to that. That way the linux-image package which is used
by the virtual meta-package does not automatically pull in
crda.

BugLink: http://bugs.launchpad.net/bugs/657901

[v2: move net/mac80211 into extra as well based on Seth's feedback]
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Seth Forshee <seth.forshee@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>