ARM: OMAP5: Reduce the static IO mapping
EMIF, GMPC and DMM driver can ioremap() the address
space as part of driver intialisation and there is
no need to have static IO mapping for them.

Hence remove the un-used static IP space and let
the respective drivers manage it as part if driver
init.

Based on the changes that was earlier done for OMAP4
which went as a part of OMAP2+ cleanup series.

 http://www.digipedia.pl/usenet/thread/19013/35491/
 by Santosh Shilimkar <santosh.shilimkar@ti.com>

Signed-off-by: R Sricharan <r.sricharan@ti.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>