ktl/check_component.py: move lbm check to linux abi function
We always have some new lbm packages that didn't came out in 'release'
pocket. And they must be on main. Thus we should ensure that the default
component for them is 'main'. So if we don't find a binary package from
lbm in the 'release' pocket, it should be on 'main'.

Now that we use linux_abi_component instead of override_component for
the component checking in the case of lbm packages, move the lbm
component special case checking to linux_abi_component function.

Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>