ktl/check_component.py: add override entry for updates-modules-*-lpia-di
In hardy, updates-modules-2.6.24-ABI-lpia-di was released originally on
universe. Thus the bot complains when it lands on main. But archive
admins just want to have it on main for the overrides.

This change adds support to use override entries for package names which
contains abis, and adds the needed entry for updates-modules in question.

Reference: http://bugs.launchpad.net/bugs/1020100
Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>