UBUNTU: Add compat-wireless v3.4
BugLink: http://bugs.launchpad.net/bugs/1023566

Add the initial compat-wireless v3.4 package.  This is based on the
latest v3.4.4 stable kernel.  This was manually created following the
steps documented at:

  http://linuxwireless.org/en/users/Download/hacking#Generating_stable_releases

  ie cd compat-wireless && ./scripts/gen-stable-release.sh -s -n -p
     Note: depends on quilt being installed

Manual creation was done because the compat-wireless archive did not
appear to have a final v3.4 based tarball, the latest available was
v3.4-rc3:

  http://www.orbit-lab.org/kernel/compat-wireless-3-stable/v3.4

Additionally, a specific request for the alx driver was made which
required the -p flag (ie apply the patches on the linux-next-pending
directory) when generating the tarball.

Also, in order to help make potential maintenance as well as upgrades
easier, we will only use a 2 digit version scheme, ie 3.4, in the
packaging.

Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>
Acked-by: Herton Krzesinski <herton.krzesinski@canonical.com>
Acked-by: Bbrad Figg <brad.figg@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>