maint-getabis: Add usage of in-tree getabis script
If the tree contains a getabis script, then use that to get the
ABI files. Add --no-tree-script option to enforce the old
indipendent method.

This also modifies the RunScript library function to allow stdout
and stderr to be printed while the command runs. Before the change
output would only be available as variables after the command had
finished.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>