Quit logmeter when a TEST_QUIT tag is received
Currently logmeter runs until it has read in a specified number of
samples from the Fluke meter.  Since we want to automate this in
the lab we should add a way for the client to indicate it has
completed the test.

This patch adds detection of a TEST_QUIT tag which will terminate
the sampling and shut down appropriately.

Signed-off-by: Colin Ian King <colin.king@canonical.com>