cve-autotriage: need to include the linux-overlay signature in branch signatures
When detecting changes in a branch we use the branch tip and list of
tags.  However the list of sha1s contained is also in part defined
by linux-overlay.  We need to include the signature of that as well so
that we rebuild when it changes.  Add the md5sum of the file to the
overall per branch signature file.

Signed-off-by: Andy Whitcroft <apw@canonical.com>